Address comments.

+++ tools/sst_dump_tool_imp.h
@@ -18,7 +18,7 @@ namespace ROCKSDB_NAMESPACE {
 class SstFileDumper {
  public:
   explicit SstFileDumper(const Options& options, const std::string& file_name,
-                         uint64_t readahead_size, bool verify_checksum,
+                         size_t readahead_size, bool verify_checksum,
                          bool output_hex, bool decode_blob_index);
 
   Status ReadSequential(bool print_kv, uint64_t read_num, bool has_from,
